Assign, sort & seperate the first net messages (opcodes) to their different servers (CharacterServer, WorldServer & WorldNode),...not finished. Add WorldServer & WorldNode projects (not used atm). Rename RealmServer->CharacterServer. Change ClassId type to byte.

+++ Projects/Framework/Database/Character/Entities/CharacterTemplateAction.cs
@@ -22,9 +22,9 @@ namespace Framework.Database.Character.Entities
     public class CharacterTemplateAction
     {
         [Field(PrimaryKey = true)]
-        public int ClassId { get; set; }
-        public int Action  { get; set; }
-        public byte Slot   { get; set; }
+        public byte ClassId { get; set; }
+        public int Action   { get; set; }
+        public byte Slot    { get; set; }
 
         public virtual CharacterTemplateClass CharacterTemplateClass { get; set; }
     }
