+++ Projects/Framework/Database/Character/Entities/CharacterTemplateClass.cs
@@ -23,7 +23,7 @@ namespace Framework.Database.Character.Entities
     public class CharacterTemplateClass
     {
         [Field(PrimaryKey = true)]
-        public int ClassId       { get; set; }
+        public byte ClassId      { get; set; }
         public int SetId         { get; set; }
         public byte FactionGroup { get; set; } // Alliance = 3, Horde = 5
